server_app/app_detector/requirements.txt
gunicorn
Flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7

torch
Pillow
numpy
torchvision

gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.22.2
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.0.147
pandas>=1.1.4
seaborn>=0.11.0
setuptools>=65.5.1 # Snyk vulnerability fix
